# config/swagger.yaml
---
swagger: '2.0'
info:
  title: Image Repository
  description: 'Repository for storing and searching for images'
  version: '1.0'
produces:
- application/json
host: localhost:3000
basePath: "/api"
paths:
  /getImages:
    get:
      summary: Returns all images.
      responses:
        '200':
          description: All images
    post:
      summary: Upload an image.
      parameters:
        - in: body
          name: Image
          description: The image to upload (replace string with name and url).
          schema:
            type: object
            required:
              - name
              - url
            properties:
              name:
                type: string
              url:
                type: string
      responses:
        '200':
          description: Uploaded image
  /getImages/{tags}:
    get:
      summary: Search for image by description/term.
      parameters:
        - name: tags
          in: path
          required: true
          description: Terms to search for (seperate by space)
          type: string
      responses:
        '200':
          description: All images with tag
  /getImages/searchByImage/{url}:
    get:
      summary: Search for similar images.
      parameters:
        - name: url
          in: path
          required: true
          description: Image to search for.
          type: string
      responses:
          '200':
            description: Similar images
  /getImages/getById/{id}:
    get:
      summary: Get image by exact id.
      parameters:
        - name: id
          in: path
          required: true
          description: ID of stored image
          type: string
      responses:
          '200':
            description: Return image success.
    delete:
      summary: Delete image by exact id.
      parameters:
        - name: id
          in: path
          required: true
          description: ID of stored image
          type: string
      responses:
          '200':
            description: Delete image success.
  